+++ SourceCode/Services/IModuleService.cs
@@ -10,6 +10,9 @@ namespace ModulesRegistry.Services
 {
     public interface IModuleService : IDataService<Module>
     {
+        Task<Module?> FindByIdAsync(ClaimsPrincipal? principal, int id, int owningPersionId);
         Task<IEnumerable<Module>> GetAllAsync(ClaimsPrincipal? principal);
+        Task<IEnumerable<Module>> GetForOwningPerson(ClaimsPrincipal? principal, int personId);
+        Task<(int Count, string Message, Module? Entity)> SaveAsync(ClaimsPrincipal? principal, Module entity, int owningPersonId);
     }
 }