+++ SourceCode/Services/ListboxItem.cs
@@ -6,5 +6,5 @@ using System.Threading.Tasks;
 
 namespace ModulesRegistry.Services
 {
-    public record ListboxItem(int Id, string Description);
+    public record ListboxItem(int Id, string Description) { public bool IsHidden { get; set; } }
 }